127 fix inactive clients not updating after new activity (#138)
* chore: moves sidebar to common folder

* 127 fix optimistic client activity updates

* 127 fix optimistic client activity updates

- Add shared temp entry builders and invoice/payment totals
- Merge activities and reminders in timeline server action
- Use optimistic updates with loading indicator on timeline and activity row
- Extract sales dialogs (invoice, record payment, payment history, edit payment)
- Refresh on tab focus with loading indicator
- Use shadcn Field components in reminder dialog
- Update activity and client-activity specs

+++ lib/actions/activity.ts
@@ -48,7 +48,7 @@ export async function getInactiveClients(days: number | null) {
       userId: clientsTable.userId,
       lastInvoiceDate: sql<string>`max(${invoicesTable.issueDate})`,
       activityCount: sql<number>`(
-        select count(*) from ${clientActivitiesTable}
+        select count(*)::int from ${clientActivitiesTable}
         where ${clientActivitiesTable.clientId} = ${clientsTable.id}
       )`,
     })
